+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Refactors the validation logic to use schema files
+- Package files only need to declare required properties now instead of every property
+
 ## [0.1.1] - 2021-05-05
 
 ### Added